Set up a bootstrap grid system for the page elements.
Created a format for importing test data and training data that you want. God damn that was a pain in the ass.

Started attaching that data to the grid when the user "chooses" it.

This is hackathon-quality code. Absolutely a fuster cluck. I intend to go back through adn justify it all after its done, but we will see.